+++ service/controller/clusterapi/v29/resource/tccpsubnets/delete.go
@@ -1,4 +1,4 @@
-package tccpsubnet
+package tccpsubnets
 
 import (
 	"context"
