+++ service/controller/clusterapi/v29/resource/tccpsubnets/error.go
@@ -1,4 +1,4 @@
-package tccpsubnet
+package tccpsubnets
 
 import (
 	"github.com/giantswarm/microerror"
